example/app/requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
anyio==3.4.0
async-timeout==4.0.1
authx==0.1.4
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
itsdangerous==2.0.1
motor==2.5.1
passlib==1.7.4
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.5
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
